requirements.txt
google-generativeai
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4